+++ Changelog.md
@@ -3,6 +3,7 @@
 ## Version 3.0.x
 ### Version 3.0
 * Add a new theme, "Dawn".
+* Add Distance Conversions.
 * Update the interface of the Quadratic Equation Solver.
 * Update the interface of the Triangles window.